backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
python-doctr[torch]==1.0.1
Pillow==11.1.0
